+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - ChangeLog.md
+- SparseCS vector
 
 ## [0.1.0] - 2020-01-05
 ### Changed